requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.7
Django==4.2.3
django-environ==0.10.0
django-filter==23.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.2
identify==2.5.24
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
nodeenv==1.8.0
openapi-codec==1.3.2
Pillow==10.0.0
platformdirs==3.9.1
pre-commit==3.3.3
prompt-toolkit==3.0.39
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
virtualenv==20.24.0
wcwidth==0.2.6
psycopg2==2.9.6
django-grappelli==3.0.6
